going through tests

+++ src/BucketSearchParallel.h
@@ -473,8 +473,8 @@ template <typename Traits> struct bucket_search_parallel_query {
   unsigned int *m_bucket_end;
   unsigned int m_nbuckets;
 
-  int *m_id_map_key;
-  int *m_id_map_value;
+  size_t *m_id_map_key;
+  size_t *m_id_map_value;
 
   ABORIA_HOST_DEVICE_IGNORE_WARN
   CUDA_HOST_DEVICE
@@ -487,8 +487,8 @@ template <typename Traits> struct bucket_search_parallel_query {
   CUDA_HOST_DEVICE
   raw_pointer find(const size_t id) const {
     const size_t n = number_of_particles();
-    int *last = m_id_map_key + n;
-    int *first = detail::lower_bound(m_id_map_key, last, id);
+    size_t *last = m_id_map_key + n;
+    size_t *first = detail::lower_bound(m_id_map_key, last, id);
     if ((first != last) && !(id < *first)) {
       return m_particles_begin + m_id_map_value[first - m_id_map_key];
     } else {
