+++ src/OctTree.h
@@ -675,8 +675,8 @@ template <typename Traits> struct octtree_query {
   vint2 *m_leaves_begin;
   int *m_nodes_begin;
 
-  int *m_id_map_key;
-  int *m_id_map_value;
+  size_t *m_id_map_key;
+  size_t *m_id_map_value;
 
   /*
    * functions for id mapping
@@ -685,8 +685,8 @@ template <typename Traits> struct octtree_query {
   CUDA_HOST_DEVICE
   raw_pointer find(const size_t id) const {
     const size_t n = number_of_particles();
-    int *last = m_id_map_key + n;
-    int *first = detail::lower_bound(m_id_map_key, last, id);
+    size_t *last = m_id_map_key + n;
+    size_t *first = detail::lower_bound(m_id_map_key, last, id);
     if ((first != last) && !(id < *first)) {
       return m_particles_begin + m_id_map_value[first - m_id_map_key];
     } else {
